chore: Separate event types in pre validation phase [DHIS2-19701]
