+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/tracker/imports/validation/ValidationCode.java
@@ -147,6 +147,7 @@ public enum ValidationCode {
   E1315("Status `{0}` does not allow creating data values, allow statuses: {1}."),
   E1316("Event cannot transition from status `{0}` to status `{1}`."),
   E1317("Generated by ProgramRule (`{0}`) - Attribute `{1}` is mandatory and cannot be deleted."),
+  E1318("Status `{0}` is not applicable for single events."),
 
   /* Relationship */
   E4000("Relationship: `{0}` cannot link to itself."),
