+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/imports/validation/validator/event/EventValidator.java
@@ -61,6 +61,7 @@ public class EventValidator implements Validator<TrackerBundle> {
                     new MetaValidator(),
                     new UpdatableFieldsValidator(),
                     new DataRelationsValidator(),
+                    new CategoryOptionComboValidator(),
                     securityOwnershipValidator,
                     all(
                         categoryOptValidator,
@@ -69,8 +70,9 @@ public class EventValidator implements Validator<TrackerBundle> {
                         new NoteValidator(),
                         new DataValuesValidator(optionService),
                         new StatusUpdateValidator(),
+                        new StatusValidator(),
                         new AssignedUserValidator()))),
-            field(TrackerBundle::getEvents, new RepeatedEventsValidator()));
+            field(TrackerBundle::getTrackerEvents, new RepeatedEventsValidator()));
   }
 
   @Override
