+++ dhis-2/dhis-services/dhis-service-tracker/src/test/java/org/hisp/dhis/tracker/imports/validation/validator/event/MandatoryFieldsValidatorTest.java
@@ -46,6 +46,7 @@ import org.hisp.dhis.tracker.imports.ValidationMode;
 import org.hisp.dhis.tracker.imports.bundle.TrackerBundle;
 import org.hisp.dhis.tracker.imports.domain.Event;
 import org.hisp.dhis.tracker.imports.domain.MetadataIdentifier;
+import org.hisp.dhis.tracker.imports.domain.SingleEvent;
 import org.hisp.dhis.tracker.imports.domain.TrackerDto;
 import org.hisp.dhis.tracker.imports.domain.TrackerEvent;
 import org.hisp.dhis.tracker.imports.preheat.TrackerPreheat;
@@ -88,13 +89,29 @@ class MandatoryFieldsValidatorTest {
   }
 
   @Test
-  void verifyEventValidationSuccess() {
+  void shouldPassValidationWhenSingleEventHasAllMandatoryFields() {
+    Event event =
+        SingleEvent.builder()
+            .event(UID.generate())
+            .orgUnit(MetadataIdentifier.ofUid(CodeGenerator.generateUid()))
+            .programStage(MetadataIdentifier.ofUid(CodeGenerator.generateUid()))
+            .program(MetadataIdentifier.ofUid(CodeGenerator.generateUid()))
+            .build();
+
+    validator.validate(reporter, bundle, event);
+
+    assertIsEmpty(reporter.getErrors());
+  }
+
+  @Test
+  void shouldPassValidationWhenTrackerEventHasAllMandatoryFields() {
     Event event =
         TrackerEvent.builder()
             .event(UID.generate())
             .orgUnit(MetadataIdentifier.ofUid(CodeGenerator.generateUid()))
             .programStage(MetadataIdentifier.ofUid(CodeGenerator.generateUid()))
             .program(MetadataIdentifier.ofUid(CodeGenerator.generateUid()))
+            .enrollment(UID.generate())
             .build();
 
     validator.validate(reporter, bundle, event);
@@ -118,12 +135,13 @@ class MandatoryFieldsValidatorTest {
   }
 
   @Test
-  void verifyEventValidationFailsOnMissingProgramStageReferenceToProgram() {
+  void shouldFailValidationWhenEventIsMissingProgramStageReferenceToProgram() {
     Event event =
         TrackerEvent.builder()
             .event(UID.generate())
             .orgUnit(MetadataIdentifier.ofUid(CodeGenerator.generateUid()))
             .programStage(MetadataIdentifier.ofUid(CodeGenerator.generateUid()))
+            .enrollment(UID.generate())
             .build();
     ProgramStage programStage = new ProgramStage();
     programStage.setUid(event.getProgramStage().getIdentifier());
@@ -149,6 +167,22 @@ class MandatoryFieldsValidatorTest {
     validator.validate(reporter, bundle, event);
 
     assertMissingProperty(reporter, event, "programStage");
+  }
+
+  @Test
+  void shouldFailValidatiWhenTrackerEventIsMissingEnrollment() {
+    Event event =
+        TrackerEvent.builder()
+            .event(UID.generate())
+            .orgUnit(MetadataIdentifier.ofUid(CodeGenerator.generateUid()))
+            .programStage(MetadataIdentifier.ofUid(CodeGenerator.generateUid()))
+            .program(MetadataIdentifier.ofUid(CodeGenerator.generateUid()))
+            .enrollment(null)
+            .build();
+
+    validator.validate(reporter, bundle, event);
+
+    assertMissingProperty(reporter, event, "enrollment");
   }
 
   @Test
