+++ dhis-2/dhis-services/dhis-service-tracker/src/test/java/org/hisp/dhis/tracker/imports/validation/validator/event/RepeatedEventsValidatorTest.java
@@ -49,6 +49,7 @@ import org.hisp.dhis.tracker.imports.TrackerImportStrategy;
 import org.hisp.dhis.tracker.imports.bundle.TrackerBundle;
 import org.hisp.dhis.tracker.imports.domain.Event;
 import org.hisp.dhis.tracker.imports.domain.MetadataIdentifier;
+import org.hisp.dhis.tracker.imports.domain.SingleEvent;
 import org.hisp.dhis.tracker.imports.domain.TrackerEvent;
 import org.hisp.dhis.tracker.imports.preheat.TrackerPreheat;
 import org.hisp.dhis.tracker.imports.validation.Error;
@@ -106,7 +107,7 @@ class RepeatedEventsValidatorTest extends TestBase {
     bundle.setEvents(events);
     events.forEach(e -> bundle.setStrategy(e, TrackerImportStrategy.CREATE_AND_UPDATE));
 
-    validator.validate(reporter, bundle, bundle.getEvents());
+    validator.validate(reporter, bundle, bundle.getTrackerEvents());
 
     assertIsEmpty(reporter.getErrors());
   }
@@ -129,7 +130,7 @@ class RepeatedEventsValidatorTest extends TestBase {
         .thenReturn(true);
     bundle.setEvents(List.of(event));
 
-    validator.validate(reporter, bundle, bundle.getEvents());
+    validator.validate(reporter, bundle, bundle.getTrackerEvents());
 
     // then
     assertHasError(
@@ -150,7 +151,7 @@ class RepeatedEventsValidatorTest extends TestBase {
     bundle.setEvents(events);
     events.forEach(e -> bundle.setStrategy(e, TrackerImportStrategy.CREATE_AND_UPDATE));
 
-    validator.validate(reporter, bundle, bundle.getEvents());
+    validator.validate(reporter, bundle, bundle.getTrackerEvents());
 
     assertAll(
         () ->
@@ -180,7 +181,7 @@ class RepeatedEventsValidatorTest extends TestBase {
     bundle.setEvents(events);
     events.forEach(e -> bundle.setStrategy(e, TrackerImportStrategy.CREATE_AND_UPDATE));
 
-    validator.validate(reporter, bundle, bundle.getEvents());
+    validator.validate(reporter, bundle, bundle.getTrackerEvents());
 
     assertIsEmpty(reporter.getErrors());
   }
@@ -197,7 +198,7 @@ class RepeatedEventsValidatorTest extends TestBase {
     reporter.addError(
         new Error("", E9999, invalidEvent.getTrackerType(), invalidEvent.getUid(), List.of()));
 
-    validator.validate(reporter, bundle, bundle.getEvents());
+    validator.validate(reporter, bundle, bundle.getTrackerEvents());
 
     assertFalse(reporter.hasErrorReport(e -> E1039 == e.getErrorCode()));
   }
@@ -214,23 +215,20 @@ class RepeatedEventsValidatorTest extends TestBase {
     bundle.setEvents(events);
     events.forEach(e -> bundle.setStrategy(e, TrackerImportStrategy.CREATE_AND_UPDATE));
 
-    validator.validate(reporter, bundle, bundle.getEvents());
+    validator.validate(reporter, bundle, bundle.getTrackerEvents());
 
     assertIsEmpty(reporter.getErrors());
   }
 
   @Test
   void testTwoSingleEventsInSameProgramStageArePassingValidation() {
-    when(preheat.getProgramStage(
-            MetadataIdentifier.ofUid(NOT_REPEATABLE_PROGRAM_STAGE_WITHOUT_REGISTRATION)))
-        .thenReturn(notRepeatebleProgramStageWithoutRegistration());
     Event eventProgramA = singleEvent();
     Event eventProgramB = singleEvent();
     List<Event> events = List.of(eventProgramA, eventProgramB);
     bundle.setEvents(events);
     events.forEach(e -> bundle.setStrategy(e, TrackerImportStrategy.CREATE_AND_UPDATE));
 
-    validator.validate(reporter, bundle, bundle.getEvents());
+    validator.validate(reporter, bundle, bundle.getTrackerEvents());
 
     assertIsEmpty(reporter.getErrors());
   }
@@ -249,27 +247,14 @@ class RepeatedEventsValidatorTest extends TestBase {
     return programStage;
   }
 
-  private ProgramStage notRepeatebleProgramStageWithoutRegistration() {
-    ProgramStage programStage = createProgramStage('A', 1, false);
-    programStage.setUid(NOT_REPEATABLE_PROGRAM_STAGE_WITHOUT_REGISTRATION);
-    programStage.setProgram(programWithoutRegistration());
-    return programStage;
-  }
-
   private Program programWithRegistration() {
     Program program = createProgram('A');
     program.setProgramType(ProgramType.WITH_REGISTRATION);
     return program;
   }
 
-  private Program programWithoutRegistration() {
-    Program program = createProgram('B');
-    program.setProgramType(ProgramType.WITHOUT_REGISTRATION);
-    return program;
-  }
-
-  private Event singleEvent() {
-    return TrackerEvent.builder()
+  private SingleEvent singleEvent() {
+    return SingleEvent.builder()
         .event(UID.generate())
         .enrollment(ENROLLMENT_B)
         .programStage(MetadataIdentifier.ofUid(NOT_REPEATABLE_PROGRAM_STAGE_WITHOUT_REGISTRATION))
