+++ dhis-2/dhis-test-e2e/src/test/java/org/hisp/dhis/tracker/imports/events/EventValidationTests.java
@@ -84,7 +84,7 @@ public class EventValidationTests extends TrackerApiTest {
 
   private static Stream<Arguments> provideValidationArguments() {
     return Stream.of(
-        Arguments.of(OU_ID, trackerProgramId, trackerProgramStageId, "E1033"),
+        Arguments.of(OU_ID, trackerProgramId, trackerProgramStageId, "E1123"),
         Arguments.arguments(null, eventProgramId, eventProgramStageId, "E1123"),
         Arguments.arguments(ouIdWithoutAccess, eventProgramId, eventProgramStageId, "E1029"),
         Arguments.arguments(OU_ID, trackerProgramId, null, "E1123"),
